requirements.txt
Flask
Pillow
Numpy